pyproject.toml
[tool.poetry]
name = "ptb"
version = "0.1.0"
description = ""
authors = ["IlyaBulatau <ilyabulatau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["all"], version = "^20.6"}
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
